fix: set the timer
configure timer activation and
deactivation of the 'resend_code' button
immediately after pressing the
'send confirmation code' button
